yosi

+++ src/diary/views.py
@@ -437,9 +437,11 @@ def month_graph(request,selected_date=None):
         # 総評コメントを取得
         month_ai=Month_AI.objects.get(user = request.user,created_date__year = year,created_date__month = month)
         ai_comment = month_ai.ai_comment
-    else:
+    elif month_ai:
       month_ai=Month_AI.objects.get(user = request.user,created_date__year = year,created_date__month = month)
       month_ai.delete()
+      ai_comment = '15日以上日記をかいてくにゃさい'
+    else:
       ai_comment = '15日以上日記をかいてくにゃさい'
     data = chart_data(emotion)
     chart_data_json = JsonResponse(data, safe=False).content.decode('utf-8')
